Convert non-ascii comments from mac-Roman or Windows-1252 as relevant, to UTF-8
Convert comments (such as "±5°") from Windows-1252 to UTF-8

Convert comments (such as "²") from Windows-1252 to UTF-8

Convert comments (such as "²" and "³") from Windows-1252 to UTF-8

Instead of Windows-1252, some of the comments seems to be in macintosh encoding.

Comments in src/MathUtils.cpp, src/TMTParser.cpp, src/TTEngine.cpp seems
to be in macintosh encoding. Comments src/TTGenerator.cpp are annoyingly
partly in macintosh encoding and partly in Windows-1252!