Replace %li with %i in format specifier in TTEngine.cpp, to avoid implicit casting to long
This is a variation of the issue with 32-bit long on Windows. On 64-bit
Linux, these %li causes casting to 64-bit types. A negative number
becomes a large positive value, 4294967296 + number, and generates
"Number too large to be parsed, larger than 32,767" errors, when used with
the "-a compile everything" option.

%li in other files than TTEngine.cpp applies to user messages, instead,
and for now, left unchanged.

Part of https://github.com/microsoft/VisualTrueType/issues/20