Use different wide string format specifier, depending on the compiler used
Reference:

https://docs.microsoft.com/en-us/cpp/c-runtime-library/format-specification-syntax-printf-and-wprintf-functions?view=msvc-160

"
The arguments that follow the format string are interpreted according
to the corresponding type character and the optional size
prefix. Conversions for character types char and wchar_t are specified
by using c or C, and single-byte and multi-byte or wide character
strings are specified by using s or S, depending on which formatting
function is being used. Character and string arguments that are
specified by using c and s are interpreted as char and char* by printf
family functions, or as wchar_t and wchar_t* by wprintf family
functions. Character and string arguments that are specified by using
C and S are interpreted as wchar_t and wchar_t* by printf family
functions, or as char and char* by wprintf family functions. This
behavior is Microsoft-specific.

Microsoft-specific:
The Z type character, and the behavior of the c, C, s, and S type
characters when they're used with the printf and wprintf functions,
are Microsoft extensions. The ISO C standard uses c and s consistently
for narrow characters and strings, and C and S for wide characters and
strings, in all formatting functions.
"

This is the truncated string problem seen in
https://github.com/microsoft/VisualTrueType/issues/20